ci-operator multi-stage: split file
